RPM Control System: Add support for multiple RPM's
This change adds support for multiple RPM's in the same location. If an
RPM request fails i.e. -rpm1 and -rpm2 (or the next number sequentially)
exists, it will try that RPM as well.

Note this does not support hydra setup's.

BUG=chromium:288433
TEST=local setup, sent a request for a hostname that exists on RPM2 and
checked that it succeeded.

Change-Id: I24c97439a8fbd9416f90432aa5a03ed8efb0b025
Reviewed-on: https://chromium-review.googlesource.com/175983
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>