[autotest] Reduce flakiness of common_util_test:testRunCommandRetrySleep
This test is known to be flaky because it relies on sleep(). It
records the time spent on RunCommand retries and checks it
against the epxected time.

This CL tries to reduce flakiness by increase the tolerated time
mismatch. If we still see it flakes, we should consider to eleminate this
test.

BUG=chromium:306534
TEST=run this test.

Change-Id: I5d712d0aea1bc6cfcd009a584e6963aa9272d83e
Reviewed-on: https://chromium-review.googlesource.com/175960
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>