Add new p2p_ServeFiles test.
This test checks that files announced from the p2p-server are served
from the p2p-http-server daemon. Note that localhost requests are
blocked.

In order to test this, a new method to wait for conditions on the
SimulatorThread is added to the lansim framework with its respective
unit test.

BUG=chromium:311657
TEST=lansim unittests and test_that p2p_ServeFiles
CQ-DEPEND=CL:176161

Change-Id: I7ff53d494240500a0a5c1ea4bbbfd73a2cc7193e
Reviewed-on: https://chromium-review.googlesource.com/176147
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>