Create a control file just for the bootperf script.
This updates the 'bootperf' script to depend on a custom-made
control file for platform_BootPerfServer.  The intent is to make it
easy to change test behaviors that 'bootperf' depends on without
interfering with suites or other clients of the boot performance
test.  This also allows documenting bootperf's dependencies.

BUG=None
TEST=run bootperf

Change-Id: Id1085e948c86db6dcafc36ce846d96c409b23dff
Reviewed-on: https://chromium-review.googlesource.com/176260
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>