Add AFDOGenerate test to generate AFDO profile data
This runs a pre-determined set of benchmarks on the
DUT under the monitoring of the "perf" linux tool and
copies the resulting data to Google Storage to be
used by the AFDO optimized build.

As part of this, modified the telemetry_runner code
to support running tests without a devserver.

BUG=None
TEST=Tested by hand using the --local flag.
     Tested with my own instance of autotest.

Change-Id: Ia237b393ab0f37a1f6e56e857c4aa39474bf72a6
Reviewed-on: https://chromium-review.googlesource.com/176286
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>