autotest: Detect ambiguous hostapd SSID references
LinuxRouter.get_ssid() will default to using the first instance of
hostapd that was configured, but this may not be desirable in those
rare cases where we configure multiple instances.  When a developer
makes an ambiguous reference, detect this and raise an error rather
than behaving in an unexpected way.

TEST=wifi_matfunc runs correctly with this change after the remaining
WiFiManager tests are migrated.
BUG=None

CQ-DEPEND=CL:176149
CQ-DEPEND=CL:176403

Change-Id: I77ffa7063da998d89ebbd2d54480f203cb67d1f5
Reviewed-on: https://chromium-review.googlesource.com/176466
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>