Bluetooth: rename BluetoothClient to BluetoothDevice
The BluetoothClient class, and BluetoothClientXmlRpcServer, were named
to match the WiFiClient class that they were based on. However for
Bluetooth tests, the DUT can actually behave as either a Client or a
Server depending on the requirements of the test - with the Tester
likewise behaving as a Server or Client appropriately.

This becomes confusing, so instead rename BluetoothClient to
BluetoothDevice so it's clear it refers to the DUT rather than its role.

BUG=chromium:256771
TEST=test_that suite:bluetooth_sanity

Change-Id: I75864de27325f3a74394a092ce7e538d8161f8a8
Reviewed-on: https://chromium-review.googlesource.com/176468
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>