[autotest] Add a long-running script that can do stat submission.
This adds a script which spawns functions located in the stats/
directory as threads so that one stat can't block another from running.

New stats can be placed in stats/ and registered with the registry, and
then they'll get automatically picked up to start running.  You can also
specify a set of servers that you'd like to be run against, and one copy
of the function will be started per server.

TEST=ran locally
BUG=chromium:340454

Change-Id: Ib548e80c46b6024b83f7307c41319b36b2d6cd35
Reviewed-on: https://chromium-review.googlesource.com/176469
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>