[autotest] Modify option no_wait for a suite job to fire and forget
no_wait is an option in run_suite for the script to exit without waiting for
the suite job to finish.

To prevent a suite job from holding on a process slot in drone for a long time.
The usage of no_wait option is expended so that the suite job can exit once all
test jobs are created, without waiting for all test jobs to finish.

related CL for test_suites:
https://chromium-review.googlesource.com/#/c/177309/

BUG=chromium:316990
TEST=trybot build: trybot-lumpy-paladin/R33-4997.0.0-b1385
then run_suite in local setup:
./run_suite.py --build=lumpy-release/R33-4933.0.0  --board=lumpy
--suite_name=dummy  -p bvt --no_wait=True
Confirm the suite job is completed right after the test jobs are scheduled.

DEPLOY=apache

Change-Id: I039a50371f9c5ebe8c8e155a9aabfb37999be4e2
Reviewed-on: https://chromium-review.googlesource.com/176601
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>