[autotest] Slow down the polling interval of run_suite.
We're currently servicing 200 get_jobs requests/second.  There used to
be a local commit on one of the servers that bumped run_suite up from 1
seconds to 5, but that's been lost over time.  Even if we go through
with the infrastructure changes to allow suites to not sit around and
poll, having run_suite check every 1 second seems a bit excessive, so
bump it to 10.

TEST=pylint
BUG=None

Change-Id: Ife0d95c6112e3537daa7ff372d2162fc3075fa5a
Reviewed-on: https://chromium-review.googlesource.com/176986
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>