[autotest] Don't mark a host as ready if it failed provision.
Currently we mark a host as ready at the end of provisioning, if
we don't call the hqes on pending. This has the following shortcomings:

1. Even though a repair is queued up against the DUT it won't be
effective till the end of the next tick in handle_agents. This
means the the DUT will join the available hosts pool for the duration
of one tick, and probably get reassigned to an hqe (might even be to
the same hqe). This is undesirable, because we know that the DUT will
switch to having a Repairing status at the end of the same tick.

2. If we have repaired a DUT too many times after a provision, we
would like to switch that DUT to repair failed, this won't happen.

This cl leaves the DUT in provisioning till the next job changes its
state, if the provisioning attempt failed, unless it was a provision
attempt without an hqe.

TEST=Ran suites where provisioning failed.
BUG=chromium:318461

Change-Id: I208287f8e59a241b8e4f6257bca8b2fba23346d2
Reviewed-on: https://chromium-review.googlesource.com/177159
Reviewed-by: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>