Add router method for adding a connected peer
Add a method that attaches a client to a running hostap instance.
In a typical testbed configuration, this means that one wireless
radio in the server instance connects to the other radio.  This
allows testing operations that the DUT may perform with other
devices connected to the same AP.

BUG=chromium:318166
TEST=Upcoming TDLS autotest

Change-Id: I1469d6a3e14e5fde33e53f05ff5bfd46420c6bc6
Reviewed-on: https://chromium-review.googlesource.com/177211
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>