(re-land) autotest: add some networking tests to BVT
This was originally CL:177138, but that was reverted in CL:177422.
The revert was determined to be erroneous, with the actual offending
CL being CL:177153.

Original commit message follows...

These autotests have been running as part of network_nightly, and
have had no failures in the past two weeks. Add them to BVT, so
that they can detect regressions quickly.

BUG=chromium:320709
TEST=Ran suite_enumerator.py, saw modified tests as listed for bvt

Change-Id: I321f35eb37af3cee850a59f8854cd7de9f5efd01
Reviewed-on: https://chromium-review.googlesource.com/177460
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>