Fix the hostname type and duplicate APs.
There was a typo, tack instead of rack.  This would result in the
AP never being turned on, hence the failure.

We also has two APs of the same model.  That mean the SSIDs were the
same which meant the result folders were the same.  This caused the AP
that ran second to fail to save results.

BUG=chromium:321394,chromium:321396
TEST=Manual

Change-Id: I7690abbcffb61169c7f7b54a674ec139c98cebc7
Reviewed-on: https://chromium-review.googlesource.com/177388
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>