avahi_utils: pass the custom config file to upstart.
This patch fixes a bug that was passing the AVAHI_DAEMON_CONF option
as an environment variable to upstart instead of an argument to the
"start" call. This fix also exposes the problem that the generated
custom config file for avahi-daemon during the test has spaces
around the "=" sign but the avahi-daemon doesn't support that.

BUG=chromium:321223
TEST=test_that p2p_ConsumeFiles passes and /var/log/messages shows the avahi-daemon running only on the faketap0 interface.

Change-Id: I339ff0b4add4e44ed19e7784e96e971447e5ddb5
Reviewed-on: https://chromium-review.googlesource.com/177416
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>