Switching the Touch FW test to the new Robot
The touchbotII robot wrapper has been under development for a while now
and seems to be quite functional.  As such I would like to switch the
test over to use it now.  This patch removed the old robot wrapper and
replaces it with the one for the new robot.

The current state of the robot test with the new robot is that it can do
the same tests that the old robot could.  The only way to really iron
out the kinks in it now is to start actually using it, I think.  More
tests will be added as this robot is much more flexible than the old
one, but replacing the old functionality was the first milestone here.

BUG=chromium:308162
TEST=I ran the full fw test using the new robot on a Lumpy in the lab
and everything worked.

Change-Id: Ife9bd73850511b4c131fc16d850e1798d686f662
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177481
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>