fakegudev: Load fake devices from a KeyFile.
Device descriptions for real devices are too complex to specify in an
environment variable. This CL changes the library to instead read in the
description from a file.

While fixing the test for this change, also break the test into data and code,
to make it slightly more maintainable.

BUG=chromium:318359
TEST=Run test-gudev.sh

Change-Id: I207074f94157e02cd77be8400dcb70a0adbfb78c
Reviewed-on: https://chromium-review.googlesource.com/177478
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>