firmware_TouchMTB hide scores behind a flag in summary
This patch hides the old 0.0-1.0 scores behind a flag. The scores
would be kept for a while until they are completely replaced by
the metrics.

BUG=chromium:316291
TEST= Replay the logs and run the summarizer:
$ tools/machine_replay.sh -b tests/logs/link
$ ./summary.sh -d /tmp
Check the summary in /tmp/summary/summary_xxxx/summary_xxxx.txt
Observe that it does not show the scores.

Now run the summarizer with "-s" flag to show scores.
$ ./summary.sh -d /tmp -s
Check the summary in /tmp/summary/summary_xxxx/summary_xxxx.txt
Observe that it does show the scores this time at the end of the file.

Change-Id: I0a5cb75f07ca497839d81cb8ded00eadf3dd8168
Reviewed-on: https://chromium-review.googlesource.com/177743
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>