Add autoupdate_CatchBadSignatures test.
This is a test to verify that update_engine correctly checks
signatures in the metadata hash and the update payload itself. This is
achieved by feeding updates to update_engine where the private key
used to make the signature, intentionally does not match with the
public key used for verification.

BUG=chromium:264352
TEST=Successfully passed on lumpy and daisy.

Change-Id: I9bced8430052b42beeea17db3d13e213fea12bef
Reviewed-on: https://chromium-review.googlesource.com/177631
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>