[autotest] Only yield results of the test if a test failed.
The parser is configured to insert a SERVER/CLIENT/Test entry
into tko for a given job, when the test aborts. This leads to
confusing logs in the waterfall, and confusing bugs, because
we get what looks like multiple failures for the same test.
Since we only care about the server job failures if a Test
didn't fail, only yield the test status.

TEST=Ran a suite, failed jobs, checked logs/bugs.
BUG=chromium:322065

Change-Id: I8ff84da102aadb8e55fbbbad1bba075dceb28270
Reviewed-on: https://chromium-review.googlesource.com/177826
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>