Revert "autotest: Add network3g_pseudomodem and network_ui tests to smoke, bvt."
Something about the way these suites are named is confusing the parser into thinking they're not experimental, and closing the tree. Since all the tests in this commit are experimental I'm reverting the change till the sheriff has had a chance to investigate on monday. This will only remove the tests from bvt/smoke. 

This reverts commit 2c51e0a7af5c88edc74256cf92987dfc0e2a4bb4.

Change-Id: I3578dac56de4ca385f1c59dc8ff140e35691578e
Reviewed-on: https://chromium-review.googlesource.com/177838
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>