autotest: Let set_device_enabled be a no-op when not supported
For its own reasons, when the lab runs server tests, it runs a near ToT
copy of the server code with control files and a client packages tied
to the image version.  This has caused WiFi tests not to work against
R32 images because of the additional set_device_enabled() method is
called on the XMLRPC server taken from the client package.  In R32,
and earlier, this method is not supported.  Rather than backport the
method call through time, lets just tolerate it being missing.

BUG=chromium:323355
TEST=Test still run correctly against ADB hosts and cros hosts.

Change-Id: Ib75554ac07da4db15b13ac3834f5075f4f0a8481
Reviewed-on: https://chromium-review.googlesource.com/177955
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>