[autotest] Add a new type of bug class MachineKillerBug
This cl has the following changes:
1) Adds a new class MachineKillerBug(Bug), which wraps
the information needed to report about a job killing a machine.
2) Move _link*(..) from reporting.TestFailure to a new file
reporting_utils, so that both TestFailure and MachineKillerBug can
use to generate links.
3)Instead of creating a generic bug, repair_utils now creats an
instance of MachineKillerBug and pass it to the reporter.

Machine killer issue now has a link to the job.

BUG=chromium:313812, chromium:314797
TEST=reporting_unittests.py, repair_utils_unittests.py
     File machine killer bug.
     File test failure bug.
     Call repair_utils.flag_problem_test to file a machine killer bug.

Change-Id: Ibe606f1e531e127578156bb358ea7a451ac137a7
Reviewed-on: https://chromium-review.googlesource.com/178702
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>