BluetoothTester: add support for basic SDP methods
BluetoothTester and BluetoothTesterXmlRpcDelegate classes are used to
expose tester-side bluetooth functionality (BluetoothSocket) so that it
could be accessed in autotests. BluetoothTester runs on the server and
basically represents the tester as an object for autotests.
BluetoothTesterXmlRpcDelegate runs on the tester and is generally a
proxy for BluetoothTester.

This CL adds connect() and service_search_request() methods to these
classes.

BUG=none
TEST=use added methods in autotests

Change-Id: I28671d2ad76d7a3e95f299c9622a6856a41ed941
Reviewed-on: https://chromium-review.googlesource.com/179167
Tested-by: Artem Rakhov <arakhov@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Artem Rakhov <arakhov@chromium.org>