pseudomodem: Add interactive state machines
This CL adds an option to run pseudomodem machines in interactive mode. When a
state machine is run in this mode, it exports a new testing object on a testing
interface, and waits for an |Advance| call on this testing interface before
taking each step. This is intended to allow tests fine grained control on the
progress of the state machine.

Specifically, this CL,
  (1) Extends StateMachine to support interactive mode
  (2) Adds a StateMachineFactory to increase flexibility in how state machines
      are instantiated by the Modem.
  (3) Extends the standalone mode of pseudomodem to run with some or all state
      machines in interactive mode.
  (4) Extends Client to add basic functionality to allow interacting with a
      machine from the cli.

BUG=chromium:326439
TEST=(1) Run pseudomodem in standalone mode and interactive mode with
         interactive state machines.
     (2) Run existing pseudomodem tests.
Change-Id: Ic448be4958904842a683bd441362ae0262c1c3bc
Reviewed-on: https://chromium-review.googlesource.com/179070
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>