firmware_TouchMTB: add CountTrackingIDFatFingerValidator
It is useful to distinguish the count tracking ID results between
gestures with and without fat fingers.

For gestures with fat fingers, thumb edge, or close fingers, we may
allow a small percentage of finger split/merge.

On the other hand, for the other gestures without using fat fingers,
we would impose a strict criteria on the count of tracking IDs.
There should be no finger split/merge, ghost finger, or other
related problems.

The new validator is implemented by simply inheriting
CountTrackingIDValidator.

BUG=chromium:326412
TEST=Replay a previous manual gesture set.
And observe that it would display CountTrackingIDFatFingerValidator
instead of CountTrackingIDValidator on the following gestures:
- fat_finger_move_with_resting_finger
- drag_edge_thumb
- two_close_fingers_tracking
- two_fat_fingers_tracking

Change-Id: I0258003c0ba3737759b5d8518c60f51e48740bf3
Reviewed-on: https://chromium-review.googlesource.com/179033
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>