firmware_TouchMTB: show test results in html report only when files exist
When a gesture set is replayed, it generates a new html report file
automatically. However, if some of the gestures/variations files are missing,
empty blocks are shown for those missing ones in the html report.

In replay mode, it is desirable to only show the test results of those
existing files.

BUG=chromium:326404
TEST=Perform the following steps
$ cd /usr/local/autotest/tests/firmware_TouchMTB

Remove all the files about one_finger_tracking
$ rm tests/logs/link/20130806_221321-fw_1.0.AA-robot/one_finger_tracking.*

Now replay the log
$ tools/machine_replay.sh -b tests/logs/link/

Observe that there are no empty blocks about one_finger_trackig.* in
the beginning of the html report file
/tmp/touch_firmware_report-link-fw_1.0.AA-complete-xxxx.html
where xxxx represents the current date-time.

Change-Id: I67e365a25c421317d08f3b4b1aa77697acf2d511
Reviewed-on: https://chromium-review.googlesource.com/179032
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>