Touch FW test: Never try use robot in MANUAL mode
Errors were showing up when the MANUAL portion of the test was being run
because the touchbotII robot wrapper was still trying to talk to the
robot even in manual mode.  This CL just adds a check to immediately
abort any robot controls unless the FW test is in ROBOT mode and pushes
all if the robot logic into the robot wrapper to avoid confusion in
test_flow.py.

BUG=chromium:326237
TEST=Ran the test and the errors went away.  The test worked fine too

Change-Id: Ic4fc607b799d88325d47eb7a7e2f86ba82acae0c
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/178970