pseudomodem: Improve dbus method logging
dbus-python catches exceptions raised when servicing a dbus remote call and
returns the error to the remote caller. This hides the exceptions from the user
of the python service. This CL add a decorator that logs the return value (or
error) of a dbus call before returning it to the remote caller.

BUG=chromium:326394
TEST=(1) Run pseudomodem in standalone mode.
     (2) Raise an exception in StateMachine::Step and make sure it gets logged.
         Without this CL, that would have been a silent death.
     (3) Run pseudomodem autotests.
Change-Id: I982a41ad028c1dec3066fd715aa7811b64870c19
Reviewed-on: https://chromium-review.googlesource.com/179013
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>