Report kernel crashes in debug info if we rollback after update attempt.
We dump assorted debug info if an update fails and we perform a rollback.
This includes information about kernel crashes stored on stateful, if
present.

These crashes will NOT be present if the kernel crashed early in the boot
and didn't have the file system mounted.

BUG=chromium:326305
TEST=pyflakes, ran unittests.

Change-Id: I52185ef88399840e8a907d1c05d39c56ee9c959d
Reviewed-on: https://chromium-review.googlesource.com/179254
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>