autotest: shill: improve failure message for DefaultProfileCreation
When the network_DefaultProfileCreation fails, it reports the missing
setting. We don't know whether just that specific setting was missing,
or if the entire file was empty. Help answer this question, by including
the full contents of the default profile, when the test fails.

BUG=chromium:325056
TEST=manual (see below)

Manual test
-----------
- modify network_DefaultProfileCreation by adding a bogus setting
  to EXPECTED_SETTINGS
- see that the test output includes the contents of the default
  profile

Change-Id: I1c1823475a92238b4836f6d357f5828bbbdfcdb2
Reviewed-on: https://chromium-review.googlesource.com/179366
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>