cryptohome.py: make CryptohomeProxy async
Move the proxy from sync to async calls which requires
signal handling.

Supporting the asynchronous methods ensures testing matches
behavior of the in use interfaces (async) and not the legacy
synchronous interfaces.  In particular, this is needed for
catching and analyzing blocking behavior on slower systems
(cpu or TPM).  There is substantial testing flake that cannot
easily be caught without aborting cryptohome on occurrence.
The support for aborting landed in prior CLs and this one
ensures we don't just see slow synchronous calls.

BUG=chromium:303677
TEST=tests pass; still catches issues if cryptohome STOP'd.

Change-Id: Ibe233a00dcd5b6f2816a370051f579b75e7be5e5
Reviewed-on: https://chromium-review.googlesource.com/179870
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>