autotest: Remove network_WiFi_SuspendTwice from wifi_matfunc
When this test fails, it usually fails because the device does not come
back out of suspend.  This is problematic because we have a small number
of WiFi cells, and no servos on the DUTs.  We should enable this test
only when we have servos to recover from suspends.

BUG=None
TEST=None

Change-Id: I3592d82c37d102e3bfe3709e07a4749468054022
Reviewed-on: https://chromium-review.googlesource.com/180033
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>