[autotest] profilers should only be installed on hosts that are DUT
Previously profilers.py tried to install profiler client on
all host in self.job.hosts. This list include all hosts
that are created on the fly, like a ServoHost.

It should really only install profiler client on the machines
passed in by option -m of autoserv.

This CL add a test checking whether a host instance is
also in self.job.machines, if not, it will be ignored.

TEST=1.Create an SSHHost(DUT_IP1) in telemetry_AFDOGenerate/control
2. /usr/bin/test_that --debug --board=lumpy DUT_IP2
--args="ignore_failures=True local=True gs_test_location=True"
telemetry_AFDOGenerate
Confirmed that the profiler client is not installed
on the DUT_IP1.
3. Remove the change, confirm that it will try to install
profiler clients on both DUT_IP1 and DUT_IP2
BUG=chromium:328507

Change-Id: Id2bc73f02079acc48b655726314aecece29fbcc0
Reviewed-on: https://chromium-review.googlesource.com/180105
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>