p2p_ServeFiles: Make active request of a p2p-server.
The test p2p_ServeFiles waits until the DUT shares a file on the p2p
network and tries to download it. It does this by scanning the
network passively until it detects the p2p-server running on the DUT
and the file shared there. Due to timing issues, the test may not see
the p2p-server starting and the DUT being announced, which is covered
by the p2p_ShareFiles tests. To avoid this situation, this patch to
p2p_ServeFiles makes an active request of the shared files on the
network instead of passively wait for the results. This active
behavior is similar to what the p2p-client implementation does.

TEST=test_that --fast -b link IP_ADDRESS  p2p_ServeFiles
BUG=chromium:328078

Change-Id: I46f8478436e2115ec4f736fd0d989895767c3d56
Reviewed-on: https://chromium-review.googlesource.com/180341
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>