[autotest] Give servo hosts the ability to update.
Since we're now staging the latest beaglebone build on the devservers
we can update all servo_hosts that are running chromeos via update engine.
This cl forces the beaglebone to check for new updates on each reset/verify,
and if one is available, download and apply it.

TEST=Update a beaglebone with and without servo(v2).
     Confirmed servod came up after the update.
BUG=chromium:272354
CQ-DEPEND=CL:178534

Change-Id: I09d0b85b81f0f60318c02e44f42b2daabe1cead4
Reviewed-on: https://chromium-review.googlesource.com/180562
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>