firmware_TouchMTB: suppport StationaryTapValidator
It is desirable to split the StationaryFingerValidator into two
validators. The first one, StationaryFingerValidator, verifies
the pulling effect of another finger on the stationary finger.
The other one, StationaryTapValidator, checks the wobble of
taps/clicks. In this way, when the stationary validator failed,
it is very easy to trace the root cause of the failure. Another
advantage is that it is possible to assign two distinct criteria
to the two validators since they are now separated.

BUG=chromium:330108
TEST=All unit tests, including the new one StationaryTapValidatorTest,
should pass.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ python tests/run_all_unittests.py

Change-Id: Ib79af6d6fcb965bc5807487454246e7841925003
Reviewed-on: https://chromium-review.googlesource.com/181092
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>