Move wifi_client creation into ap_runner; refactor
Now that wifi_client.close() has been fixed it can be moved
back into the runner.  When the ability to clear logs and rebooting
the DUT in the runner it will be cleaner to have the wifi_client
defined in the runner.

The RPC server takes about 20 seconds for bring up and tear down.
With 130 APs this can take 45 more minutes if it is done per AP.
Instead we will reset the RPC server per batch, which is for every
15 APs.  That only adds 3 minutes to a multi-hour test.

This CL also moves all of the SSID verification into its own method.
This makes the code a lot more readable and it releases APs back into
the pool faster.  Previously APs that failed to configure were locked
until the entire test ended.  Now they are released as soon as a
misconfiguration is discovered.

TEST=Manual; run control.debug
BUG=None

Change-Id: I7df36e6c8307c7673d7b5ccbc4db5ff05b851f5e
Reviewed-on: https://chromium-review.googlesource.com/181693
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>