[autotest] Add faulthandler to autoserv
This change adds the faulthandler package as a new external_package.
It then imports it to autoserv and then registers SIGTERM. This will
dump the thread's current traceback on SIGTERM then execute the
previous signal handler.

BUG=chromium:302815
DEPLOY=build_externals
TEST=Executed build_externals and ensured faulthandler was downloaded
and installed and that re-executing build_externals did not reinstall
it. Then ran sleeptest with a 1 minute timeout and ensured the
faulthandler output showed up in autoserv.DEBUG when the scheduler
aborted it.

Change-Id: Id59f897bb766a821ef463c8166531db480c68041
Reviewed-on: https://chromium-review.googlesource.com/181688
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>